requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
psutil==5.6.3
PyYAML==5.1.2
uvloop==0.13.0
yarl==1.3.0
